+++ lib/l10n/editor_l10n.dart
@@ -137,6 +137,8 @@ class EditorL10n {
       locale.languageCode == 'zh' ? '连线时复制' : 'Copy on Connect';
   String get collapseExpand =>
       locale.languageCode == 'zh' ? '折叠/展开控件' : 'Collapse/Expand Controls';
+  String get collapsible =>
+      locale.languageCode == 'zh' ? '可折叠' : 'Collapsible';
   String get lock => _t('lock');
   String get unlock => _t('unlock');
   String get lockShortcut => _t('lockShortcut');